% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{time_treat}
\alias{time_treat}
\title{Time the Seam-Carving image reduction of an image}
\usage{
time_treat(in_vec, height, width, channels, verbose)
}
\arguments{
\item{in_vec}{Rcpp::NumericVector input vector, with dimension c(height * width * channels)}

\item{height}{int number of rows in the image}

\item{width}{int number of columns in the image}

\item{channels}{int number of byte per pixel in the image}

\item{verbose}{bool set to true to print progress to console}
}
\description{
Seam-Carving with Dijkstra path-finding

This is the C++ implementation, you can use the R convenience function timedrun( img, verbose )
}
