% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seamcarvdijkstra.R
\name{timedrun}
\alias{timedrun}
\title{Time the Seam-Carving image reduction of an image}
\usage{
timedrun(img, verbose)
}
\arguments{
\item{img}{an array, with dimension c(height, width, channels), as returned by recolorize.readImage}

\item{verbose}{a boolean, set to TRUE to print progress to console}
}
\value{
the computing time, in milliseconds
}
\description{
time Seam-Carving with Dijkstra path-finding
}
