% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seamcarvdijkstra.R
\name{treat}
\alias{treat}
\title{Apply Seam-Carving image reduction to an image}
\usage{
treat(img, verbose)
}
\arguments{
\item{img}{an array, with dimension c(height, width, channels), as returned by recolorize.readImage}

\item{verbose}{a boolean, set to TRUE to print progress to console}
}
\value{
an array, with dimension c(height, width - 1, channels), containing the reduced image
}
\description{
Seam-Carving with Dijkstra path-finding
}
