% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{treat_img}
\alias{treat_img}
\title{Apply Seam-Carving image reduction to an image}
\usage{
treat_img(in_vec, height, width, channels, verbose)
}
\arguments{
\item{in_vec}{Rcpp::NumericVector input vector, with dimension c(height * width * channels)}

\item{height}{int number of rows in the image}

\item{width}{int number of columns in the image}

\item{channels}{int number of byte per pixel in the image}

\item{verbose}{bool set to true to print progress to console}
}
\description{
Seam-Carving with Dijkstra path-finding

This is the C++ implementation, you can use the R convenience function treat( img, verbose )
}
